+++ packages/design-system-docs/src/pages/components/Dialog/_Dialog.docs.scss
@@ -28,7 +28,7 @@ Style guide: components.dialog.size
 
 @react-example Dialog.example.jsx
 
-@react-props Dialog.jsx
+@react-props Dialog.tsx
 
 Style guide: components.dialog.react
 */
