+++ packages/design-system-docs/src/pages/components/Dialog/_ThirdPartyLink.docs.scss
@@ -13,7 +13,7 @@ Style guide: patterns.external-link
 
 @react-example ThirdPartyLink.example.jsx
 
-@react-props Dialog.jsx
+@react-props Dialog.tsx
 
 Style guide: patterns.external-link.react
 */
