--- packages/design-system/src/types/Dialog/Dialog.d.ts
@@ -1,133 +1,0 @@
-/* eslint-disable filenames/match-exported */
-import * as React from 'react';
-
-export type DialogCloseButtonSize = 'small' | 'big';
-
-export type DialogSize = 'narrow' | 'wide' | 'full';
-
-export interface DialogProps {
-  /**
-   * If `true`, the modal will receive a role of `alertdialog`, instead of its
-   * default `dialog`. The `alertdialog` role should only be used when an
-   * alert, error, or warning occurs.
-   */
-  alert?: boolean;
-  /**
-   * Analytics events tracking is enabled by default. Set this value to `false` to disable tracking for this component instance.
-   */
-  analytics?: boolean;
-  /**
-   * An override for the dynamic content sent to analytics services. By default this content comes from the heading.
-   *
-   * In cases where this component’s heading may contain **sensitive information**, use this prop to override what is sent to analytics.
-   */
-  analyticsLabelOverride?: string;
-  /**
-   * Provide a **DOM node** which contains your page's content (which the modal should render
-   * outside of). When the modal is open this node will receive `aria-hidden="true"`.
-   * This can help screen readers understand what's going on.
-   * Also see `getApplicationNode`.
-   */
-  applicationNode?: any;
-  /**
-   * Buttons or other HTML to be rendered in the "actions" bar
-   * at the bottom of the dialog.
-   */
-  actions?: React.ReactNode;
-  /**
-   * Additional classes to be added to the actions container.
-   */
-  actionsClassName?: string;
-  /**
-   * Aria label for the close button
-   */
-  ariaCloseLabel?: string;
-  /**
-   * The modal's body content
-   */
-  children: React.ReactNode;
-  /**
-   * Additional classes to be added to the root dialog element.
-   */
-  className?: string;
-  /**
-   * Size of the close button. See [Button component]({{root}}/components/button/#components.button.react)
-   */
-  closeButtonSize?: DialogCloseButtonSize;
-  /**
-   * For internationalization purposes, the text for the "Close" button must be
-   * passed in as a prop.
-   */
-  closeButtonText?: React.ReactNode;
-  /**
-   * Variation string to be applied to close button component. See [Button component]({{root}}/components/button/#components.button.react)
-   */
-  closeButtonVariation?: string;
-  /**
-   * The icon to display as part of the close button
-   */
-  closeIcon?: React.ReactNode;
-  /**
-   * @hide-prop [Deprecated] This prop has been renamed to `closeButtonText`
-   * @hide-prop The text for the "Close" button
-   */
-  closeText?: React.ReactNode;
-  /**
-   * Enable exiting the dialog when a user presses the Escape key.
-   * [Read more on react-aria-modal docs.](https://github.com/davidtheclark/react-aria-modal#escapeexits)
-   */
-  escapeExits?: boolean;
-  /**
-   * @hide-prop [Deprecated] This prop has been renamed to `escapeExits`.
-   * @hide-prop Disable exiting the dialog when a user presses the Escape key.
-   */
-  escapeExitDisabled?: boolean;
-  /**
-   * Same as `applicationNode`, but a function that returns the node instead of
-   * the node itself. The function will not be called until after the component
-   * mounts, so it's safe to use browser globals and refer to DOM nodes within
-   * it (e.g. `document.getElementById(..)`)
-   */
-  getApplicationNode?: (...args: any[]) => any;
-  /**
-   * Additional classes to be added to the header, which wraps the heading and
-   * close button.
-   */
-  headerClassName?: string;
-  /**
-   * The Dialog's heading, to be rendered in the header alongside the close button.
-   */
-  heading?: React.ReactNode;
-  /**
-   * Set focus to a specific element that should receive initial focus (if `focusDialog={false}`).
-   * [Read more on react-aria-modal docs.](https://github.com/davidtheclark/react-aria-modal#initialfocus)
-   */
-  initialFocus?: string;
-  /**
-   * A method to handle the state change of exiting (or deactivating)
-   * the modal. It will be invoked when the user presses Escape, or clicks outside
-   * the dialog (if `underlayClickExits=true`).
-   */
-  onExit?: (...args: any[]) => any;
-  size?: DialogSize;
-  /**
-   * @hide-prop [Deprecated] This prop has been renamed to `heading`.
-   */
-  title?: React.ReactNode;
-  /**
-   * Enable exiting the dialog when a user clicks the underlay.
-   */
-  underlayClickExits?: boolean;
-  /**
-   * Allow additional AriaModal props to be passed to Dialog
-   */
-  // eslint-disable-next-line camelcase
-  [additional_props: string]: unknown;
-}
-
-type OmitProps = 'size' | 'title';
-
-// Remove the "size" definition inside React.ComponentsWithRef, and use ours instead
-declare const Dialog: React.FC<Omit<React.ComponentPropsWithRef<'div'>, OmitProps> & DialogProps>;
-
-export default Dialog;
